added react-route-dom and organise src folder
